// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Status = {
  NEW: 'NEW',
  MISSED_CALL: 'MISSED_CALL',
  NEGOTIATION: 'NEGOTIATION',
  REJECTED: 'REJECTED'
} as const

export type Status = (typeof Status)[keyof typeof Status]